Implement dao in purchase orders

+++ sales/purchase_orders_service/src/api/schemas/product_schema.py
@@ -3,5 +3,5 @@ from marshmallow import Schema, fields
 
 class ProductsSchema(Schema):
     product_id = fields.Int(required=True)
-    product_price = fields.Double(required=True)
+    product_price = fields.Float(required=True)
     product_quantity = fields.Int(required=True)
